Addition of surface timers and removal of tileset managment.
So, what I've done, is that I've removed the tileset management, and
replaced it with something more general: surface timers! These function
extremely well, as NO graphics load on startup, and they are only loaded
as and if they are needed. After 10 seconds of not being used, they'll
be unloaded, allowing you to be sparing on memory, and there is never a
noticable loading lag spike!