Fixed stupidity (#3), and other updates.
Another (Well, thing that I carelessly missed out) problem was that I
forgot the updating code in RenderTextureByRects, causing items and the
like to not actually show. I added it, after I noticed items didn't
draw.

I've also added a music overide for the IDE on FMOD, 'cause I was
testing out some music manipulation stuff. Currently, it is set to true,
so, if it's not needed for you, change the MusicOveride constant to
false!

Object-oriented code, a GameLoop overhaul, and an editor suite will be
coming your way soon! Keep an eye out!