Change of TryCreateDirectX8Device
Here, I've simply rearragned the order of the procedure
TryCreateDirectX8Device. It now tries in the order of Pure Device,
Hardware Vertex, Software Vertex, Mixed Vertex, and fail. Don't know
what Pure Device is? Read it next to the code itself!

Apologies for the massive amount of other changes. GitHub is fairly
arbitrary.